restricted automod to certain guilds

+++ index.js
@@ -4,19 +4,23 @@ const client = new Discord.Client({ partials: ['MESSAGE', 'CHANNEL', 'REACTION']
 const token = process.env.TOKEN
 const badword = process.env.BADWORDS.toLowerCase()
 const goodword = process.env.GOODWORDS.toLowerCase()
+const guilds = process.env.GUILDS.loLowerCase()
 
 client.on("message", (message) => {
   content = message.content
   channelId = message.channel.id
-  if (message.channel.type !== "dm") categoryId = message.channel.parent.id
-   else categoryId = null
-  if(content.includes("no xp role") && channelId == 848929953641791555)
-   message.react("<:noXP:848925909858910259>")
-  if (content.toLowerCase().includes(badword)
-  && !content.toLowerCase().includes(goodword) &&
-  categoryId != 847264659455082527)
-   message.delete().catch(error =>
-     console.log(error))
+  guildId = message.guild.id
+  if (guildId == guilds) {
+    if (message.channel.type !== "dm") categoryId = message.channel.parent.id
+     else categoryId = null
+    if(content.includes("no xp role") && channelId == 848929953641791555)
+     message.react("<:noXP:848925909858910259>")
+    if (content.toLowerCase().includes(badword)
+    && !content.toLowerCase().includes(goodword) &&
+    categoryId != 847264659455082527)
+     message.delete().catch(error =>
+       console.log(error))
+  }
 
 
 })
